timeseries -- add an aggregation method
I'm a little disturbed by how similar timeseries is to the storing
part of score.

The implementations are pretty different though, and for a reason.
This is meant to be fastish and small (hence sqlite), qsscore
is meant to be more feature complete and hence json
and more adhoc data.